Adds package metadata via setup.py
I based this on an example at http://guide.python-distribute.org/creation.html with your GitHub project/user info. I went out on a limb and said it requires Django 1.4 based on your claim that it works on 1.5 and likely earlier.

This file seems to be all that's necessary to `pip install` via a repo URL and should make submission to PyPI (#1) much easier for even simpler re-use of this code.